pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "douyin-video-mcp"
version = "1.0.0"
description = "MCP server for downloading Douyin videos and Using ASR API to extract text"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "fancyboi999", email = "fancyboi999@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "requests",
    "ffmpeg-python",
    "tqdm",
    "dashscope",
]

[project.urls]
Homepage = "https://github.com/fancyboi999/douyin-video-mcp"
Repository = "https://github.com/fancyboi999/douyin-video-mcp"
Issues = "https://github.com/fancyboi999/douyin-video-mcp/issues"

[project.scripts]
douyin-video-mcp = "douyin_video_mcp.__main__:main"
